refactor: main.js -> App 컴포넌트 분리
